moved the button to the inbox page
